+++ packages/core/src/useMobileLandscape/interface.ts
@@ -1,0 +1,6 @@
+/**
+ * @title useMobileLandscape
+ * @returns 是否是移动端横屏
+ * @returns_en whether is mobile landscape
+ */
+export type UseMobileLandscape = () => boolean;