+++ packages/core/changelog.md
@@ -272,3 +272,16 @@
 ### Core
 
 * fix(useWindowScroll): ssr error
+
+## 5.0.9(March 17, 2024)
+
+### Core
+
+* add usePlatform
+* add useMobileLandscape
+
+## 5.0.10(March 17, 2024)
+
+### Core
+
+* add deps in usePlatform useCallback